Researcher Inviting the Reviewer
review_collaborator.jsx searches and adds Reviewers to the Project being worked on by the Researcher. It searches and suggests by fullName.

reviewdb.jsx displays the project invitations and on confirmation the reviewer is added to the project. Since we are fetching by projectID we could easily view content of the project.